requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
groq==0.11.0
yt-dlp==2024.12.13
opencv-python==4.10.0.84
mediapipe>=0.10.0
pillow==11.0.0
moviepy==1.0.3
numpy>=1.24.0,<2.0.0
pydantic==2.9.2

